@@ -2,8 +2,9 @@
   "name": "@angular-extensions/elements",
   "version": "14.0.0",
   "peerDependencies": {
-    "@angular/common": ">=12",
-    "@angular/core": ">=12"
+    "@angular/common": ">=14",
+    "@angular/core": ">=14",
+    "rxjs": ">=7.5.0"
   },
   "dependencies": {
     "tslib": "^2.1.0"
