@@ -1,5 +1,5 @@
 import { Component, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { LazyElementDynamicDirective } from './lazy-element-dynamic.directive';
 
@@ -36,12 +36,12 @@ describe('LazyElementDirectiveDynamic', () => {
   let fixture: ComponentFixture<TestHostComponent>;
   let appendChildSpy: jest.SpyInstance;
 
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
       declarations: [TestHostComponent, LazyElementDynamicDirective],
     }).compileComponents();
-  }));
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(TestHostComponent);
